refactor(refs T30969): move DpModal to component root
