+++ src/components/core/DpEditor/DpUploadModal.vue
@@ -52,7 +52,7 @@
 
 <script>
 import DpInput from '../../DpInput/DpInput'
-import DpModal from '../DpModal'
+import DpModal from '../../DpModal/DpModal'
 import DpUploadFiles from '../DpUpload/DpUploadFiles'
 
 export default {
